move EventSourcing.Serialization.NewtonsoftJson DI code into main EventSourcing.Serialization.NewtonsoftJson package
