+++ EventSourcing.Persistence/PostgreSql/Persistence.PostgreSql.IntegrationTests/PostgreSqlCollectionFixture.cs
@@ -6,9 +6,9 @@ using DbUp;
 using EventSourcing.Bus.Abstractions;
 using EventSourcing.Extensions;
 using EventSourcing.Extensions.DatabaseMigrations.Persistence.PostgreSql.DbUp.Extensions;
-using EventSourcing.Extensions.DependencyInjection.Serialization.NewtonsoftJson;
 using EventSourcing.Persistence.PostgreSql;
 using EventSourcing.Persistence.PostgreSql.Extensions;
+using EventSourcing.Serialization.NewtonsoftJson.Extensions;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
