+++ EventSourcing.Persistence/SqlServer/Persistence.SqlServer.IntegrationTests/Persistence.SqlServer.IntegrationTests.csproj
@@ -24,7 +24,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\..\EventSourcing.Serialization\Extensions.DependencyInjection.Serialization.NewtonsoftJson\Extensions.DependencyInjection.Serialization.NewtonsoftJson.csproj" />
+      <ProjectReference Include="..\..\..\EventSourcing.Serialization\Serialization.NewtonsoftJson\Serialization.NewtonsoftJson.csproj" />
       <ProjectReference Include="..\..\Persistence.IntegrationTests.Base\Persistence.IntegrationTests.Base.csproj" />
       <ProjectReference Include="..\Extensions.DatabaseMigrations.Persistence.SqlServer.DbUp\Extensions.DatabaseMigrations.Persistence.SqlServer.DbUp.csproj" />
       <ProjectReference Include="..\Persistence.SqlServer\Persistence.SqlServer.csproj" />
