+++ EventSourcing.Persistence/SqlServer/Persistence.SqlServer.IntegrationTests/SqlServerCollectionFixture.cs
@@ -5,9 +5,9 @@ using DbUp;
 using EventSourcing.Bus.Abstractions;
 using EventSourcing.Extensions;
 using EventSourcing.Extensions.DatabaseMigrations.Persistence.SqlServer.DbUp.Extensions;
-using EventSourcing.Extensions.DependencyInjection.Serialization.NewtonsoftJson;
 using EventSourcing.Persistence.SqlServer;
 using EventSourcing.Persistence.SqlServer.Extensions;
+using EventSourcing.Serialization.NewtonsoftJson.Extensions;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
