--- EventSourcing.Serialization/Extensions.DependencyInjection.Serialization.NewtonsoftJson/Extensions.DependencyInjection.Serialization.NewtonsoftJson.csproj
@@ -1,35 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
-        <TreatWarningsAsErrors Condition=" '$(Configuration)' == 'Release' ">true</TreatWarningsAsErrors>
-        <GenerateDocumentationFile>true</GenerateDocumentationFile>
-        <RootNamespace>EventSourcing.Extensions.DependencyInjection.Serialization.NewtonsoftJson</RootNamespace>
-        <AssemblyName>EventSourcing.Extensions.DependencyInjection.Serialization.NewtonsoftJson</AssemblyName>
-    </PropertyGroup>
-
-    <!-- NuGet metadata -->
-    <PropertyGroup>
-        <Title>EventSourcing.Extensions.DependencyInjection.Serialization.NewtonsoftJson</Title>
-        <Description>This package contains DependencyInjection extensions for NewtonsoftJson implementation of EventSourcing.Serialization.Abstraction package.</Description>
-        <Authors>Dawid Szmidka</Authors>
-        <PackageId>EventSourcing.Extensions.DependencyInjection.Serialization.NewtonsoftJson</PackageId>
-        <Version>0.1.0-alpha</Version>
-        <PackageReleaseNotes>
-            - Initial version of EventSourcing.Extensions.DependencyInjection.Serialization.NewtonsoftJson project
-        </PackageReleaseNotes>
-        <PackageLicenseExpression>MIT</PackageLicenseExpression>
-        <PackageTags>EventSourcing;EventSourcing.Extensions;EventSourcing.Extensions.DependencyInjection</PackageTags>
-        <IncludeSymbols>true</IncludeSymbols>
-        <SymbolPackageFormat>snupkg</SymbolPackageFormat>
-        <IncludeSource>true</IncludeSource>
-        <RepositoryUrl>https://github.com/TheName/EventSourcing</RepositoryUrl>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\Serialization.Abstractions\Serialization.Abstractions.csproj" />
-        <ProjectReference Include="..\Serialization.NewtonsoftJson\Serialization.NewtonsoftJson.csproj" />
-        <ProjectReference Include="..\Serialization\Serialization.csproj" />
-    </ItemGroup>
-
-</Project>
