+++ EventSourcing.Serialization/Serialization.NewtonsoftJson/Extensions/EventSourcingSerializationBuilderExtensions.cs
@@ -1,11 +1,10 @@
 ﻿using System;
 using EventSourcing.Serialization.Abstractions;
 using EventSourcing.Serialization.Abstractions.DependencyInjection;
-using EventSourcing.Serialization.NewtonsoftJson;
 using Microsoft.Extensions.DependencyInjection;
 using Newtonsoft.Json;
 
-namespace EventSourcing.Extensions.DependencyInjection.Serialization.NewtonsoftJson
+namespace EventSourcing.Serialization.NewtonsoftJson.Extensions
 {
     /// <summary>
     /// The <see cref="IEventSourcingSerializationBuilder"/> extensions.
